Adds adds moves and restructures files
