+++ src/moves/dark/darkTemplate.py
