+++ src/moves/fairy/fairyTemplate.py
