+++ src/moves/fighting/fightingTemplate.py
