+++ src/moves/fire/ember.py
