+++ src/moves/fire/fireFang.py
