+++ src/moves/fire/fireSpin.py
