+++ src/moves/fire/fireTemplate.py
