+++ src/moves/fire/flamethrower.py
