+++ src/moves/fire/flareBlitz.py
