+++ src/moves/fire/inferno.py
