+++ src/moves/ghost/ghostTemplate.py
