+++ src/moves/normal/normalTemplate.py
