+++ src/moves/psychic/psychicTemplate.py
